More get_config work
Figure out what serde is converting the tracking_numbers to

+++ src/structs.rs
@@ -2,8 +2,8 @@ pub struct BoxStruct
 {
     pub x_pos: u32,
     pub y_pos: u32,
-    pub tracking_number: u32,
-    pub destination: u32
+    pub tracking_number: i64,
+    pub destination: i64
 }
 
 pub struct ConnectionStruct
